Adds support for consistency headers (#8)
The service now supports x-ms-consistency-level and x-ms-session-token HTTP headers for controlling Cosmos DB consistency. There is also a configuration setting for the default consistency level to use.

We're adding the HTTP request and response headers to the IFhirContext, and a DocumentClientWithConsistencyLevelFromContext class, which implements IDocumentClient, is responsible biding a database request's consistency settings with the IFhirContext's headers.

Most of DocumentClientWithConsistencyLevelFromContext is generated at build time (with incremental build awareness) with a tool. This tool uses Roslyn to generate the code and can be used in the future for other coge generation needs (such as an instrumented IDocumentClient for telemetry).