Supports _type search parameter, search across all resource types, and :missing modifier (#154)
or access control, we need to the resource type to be included in the search expression tree. This also opens up the possibility of supporting the _type search parameter and searches across all resource types (/?_lastUpdated=gt2010-10-01 in addition to /Observation?_lastUpdated=gt2010-10-01).

Also makes an adjustment to the search expression trees, to explicitly state which search parameter an expression tree is bound to, by introducing a SearchParameterExpression type. Previously, there was just a convention, an implicit contract between the expression parser and the data layer. This meant that (AND (AND a b) (AND c d)) was not the same as (AND a b c d), so tree rewriting would not be a safe thing to do.

Finally, we implement support for the :missing search parameter modifier.

Resolves #152
Resolves #153