Adds request charge header and makes max continuation token size configurable (#188)
The RUs consumed during a request will now be included in the response headers.
Also allowing the maximmum continuation token size to be set via config. By default, Cosmos DB's tokens can grow up to 16KB, which could be problematic for some clients since the token needs to be a URI query parameter. It can now be capped at 1KB per request. We are also doing away with the approach of storing the continuation token in a separate document in the database.

Also consolidating the two codegened IDocumentClient classes into one.