Initial export job worker. (#434)
Adding the initial export job worker code. The job worker is hosted as a background service and periodically checks to see if there is a new export job available.

To support allowing multiple workers, the logic to determine how many new jobs are available is done the stored procedure. It also handles resuming stale job in case the job process terminates unexpectedly and did not finish processing the job correctly,