Set modelinfo provider in assembly fixture (#481)
This is a temporary workaround to stabilize unit tests after the STU3 refactoring. Extends our xUnit extension to include a TestFixtureAttribute, which allows a type to be created before any test executes (and disposed at the end). There is a sample for how to create such an extension in the xUnit repo.

We then use this attribute to define a assembly fixture that sets the model info provider.
This should become unnecessary as the STU3/R4 work advances.