Move crucible tests to their own project, and split for Stu3 and R4 (#557)
This updates our E2E test projects to run the Crucible tests for both Stu3 and R4.  Initial goal was to separate out the Crucible tests to their own test project to be able to more easily run them independently.  However, the test runs were not successfull.  They are still combined and the split may be done later.

* Move crucible tests to their own project, and split for Stu3 and R4

* Separate Known Crucible tests by Stu3 and R4.

* Clean up extra comments and move string to KnownCrucible file.

* Update environment variables using environment constant.

* Undo change to split Crucible tests to their own project.

* Add test resourcetest_immunizationevaluation/X010_ImmunizationEvaluation with POCO literal issue

* Add one more crucible test failure.

* Another crucible failure.