Update sql server password to MI (#3946)
This changes the ARM deployment to use Managed Identity exclusively vs SQL admin & password. This affects new deployments only. Users that use the new ARM deployment will not be able to connect to an existing SQL server that was configured to use passwords. They would have to create a new SQL Server.