Fix Proxy URL Generation for Bundles (#5073)
Users can specify X-Forwarded-Host and X-Forwarded-Prefix to modify the returned URLs for resources in scenarios where the FHIR server is behind a proxy. However, this logic was not observed for requests within bundles. This is because the injected ForwardedHeaderMiddleware that updates the URL in the request context was running after the middleware in the IStartupFilter that generated the FHIR context; so the FHIR context was never updated properly.

This PR updates the order of middleware registration by adding the ForwardedHeaderMiddleware inside of the same IStartupFilter.