[Bundles] Fix multi parallel reference resolution (#5083)
The main change in this PR is that for parallel bundles we're not going to use ResourceIdProvider to propagate IDs created for dynamic references during transactions. As ResourceIdProvider is not thread safe, we're from now on following a different approach: IDs generated by dynamic references are propagated as part of the child inner request. ResourceIdProvider is still used to generate the IDs, just the propagation of these value will follow a different path.

The main issue is in the execution of parallel transactions with more than one generated identifier.

Change CreateResourceHandler to identify the execution being handled by a parallel bundle and rely in the persisted ID created at the bundle handler. That's the main change to avoid new IDs being created and overriding already resolved references.
Assing new values to BundleResourceContext: persisted ID and bundle type are now part of the execution.
Removing the dependency on ResourceIdProvider.Create overrides during inner requests of parallel bundles, as this is not a thread-safe solution.